gcp/cloud_functions/ingest_ministerio/requirements.txt
google-cloud-storage
requests
flask
